moved createNextChunkArrivedPromise to utils
